Rename init_gplay command to sync_gplay
